Replace GenericRepository, GenericRepositoryImpl and AbstractGenericDao with single UserRepository (JpaRepository implementation). Deleted spare @Bean declarations from WebConfig.java, replace sessionFactory bean to entityManagerFactory and refactor corresponding transactionManager.

+++ src/main/java/com/innowisegroup/simpleblog/service/UserService.java
@@ -8,10 +8,10 @@ import java.util.List;
 
 public interface UserService {
     List<UserDto> getAllUsers();
-    UserDto getUserById(int id);
+    UserDto getUserById(long id);
     void createUser(UserDto userDTO) throws UserLastnameValidationException;
-    void updateUser(int id, UserDto userDTO);
-    void deleteUser(int id);
+    void updateUser(long id, UserDto userDTO);
+    void deleteUser(long id);
     List<String> getUsersPasswords();
     List<UserDto> getSortedByNameUsersByRole(UserRole role);
     List<UserDto> getUsersWithCapitalizedNames();
